requirements.txt
langchain==1.2.6
langchain-google-genai==4.2.0
langchain-openai==1.1.7
psycopg[binary]
fastapi
uvicorn
python-multipart
pydantic
djangorestframework
djangorestframework-simplejwt
python-dotenv
pytest
pytest-asyncio